{"schema":"decodex/commit/1","summary":"Split ingestion profile admin modules","authority":"manual"}
